+++ apps/web/app/(authentication)/login/page.tsx
@@ -6,9 +6,9 @@ import Logo from "@/components/shared/logo";
 const Page = () => {
   return (
     <div className="flex h-screen w-screen items-center justify-center">
-      <div className="flex w-full max-w-sm flex-col items-center justify-center gap-6 px-4">
+      <div className="flex flex-col items-center justify-center gap-6 w-full max-w-sm px-4">
         <Logo />
-        <div>
+        <div >
           <SocialButton />
         </div>
       </div>
